fix(OMN-263): parent-PID fast path, consolidated SERVER_PATH, documented lock-substring breadth
Third /code-review high pass on PR #220 confirmed four findings; this
addresses all of them:

1. (correctness) The self-PID fast path's rationale misread the repo's
   process topology: globalSetup writes the vitest MAIN process's pid
   to the lock, but per-file afterAll fullCleanup() calls run in the
   forked WORKER — a different pid — so the shortcut never fired there
   and every teardown still shelled out to `ps`. The worker's direct
   parent IS the main process (tinypool forks workers from it), so
   holderPid === process.ppid is an exact identity for that case, not
   a heuristic. Added it (guarded against ppid 1 for orphaned,
   reparented workers), corrected the docstring, and added a
   regression test that exercises the real worker→main relationship.

2. (correctness, documented + ticketed rather than patched) The lock's
   'vitest' substring can falsely confirm identity for any unrelated
   vitest process that reuses a dead holder's PID. It cannot be made
   checkout-specific like SERVER_PATH: the lock is machine-global by
   design (one OmniFocus per machine), so a legitimate holder can be a
   run from ANY worktree — under-matching would steal a live holder's
   lock (the dangerous 2026-06-09 direction), while over-matching only
   causes a false refusal with an actionable message (the safe
   direction). The exact fix — recording process start time in the
   lock file — is a lock-format change tracked in OMN-265; the
   deliberate-breadth decision is now documented at the constant.

3. (cleanup) SERVER_PATH existed twice — mcp-test-client.ts (via
   import.meta.url) and unified-test-server.ts (via __dirname) —
   computed differently for the same value. Consolidated into
   tests/integration/helpers/server-path.ts; both spawners and
   shared-server.ts's identity check now import the one constant, so
   the spawn path and the identity check can never drift apart.

4. (cleanup) The confirmed-mismatch-vs-unverifiable pattern at three
   sites now cross-references its siblings, with an explicit warning
   that the unverifiable fallbacks deliberately point in different
   directions (refuse for the lock, proceed-with-kill for the shared
   server) and must not be blindly unified.
