+++ tests/unit/support/integration-guard-self-pid.test.ts
@@ -1,10 +1,11 @@
-// OMN-263 code-review follow-up: proves the self-PID shortcut in
-// integration-guard.ts's default verifyPidIdentity actually avoids shelling
-// out to `ps` for the common case (checking your own currently-held lock),
-// rather than just asserting the end result. Isolated into its own file
-// because it mocks `child_process` module-wide — integration-guard.test.ts
-// has tests that deliberately exercise the real `ps` binary and would break
-// under this mock.
+// OMN-263 code-review follow-up: proves the self-PID and parent-PID
+// shortcuts in integration-guard.ts's default verifyPidIdentity actually
+// avoid shelling out to `ps` for the in-run cases (the main process checking
+// its own lock; the forked worker checking the main's lock — the worker's
+// direct parent), rather than just asserting the end result. Isolated into
+// its own file because it mocks `child_process` module-wide —
+// integration-guard.test.ts has tests that deliberately exercise the real
+// `ps` binary and would break under this mock.
 //
 // getProcessCommand (in integration-guard.ts) is the ONLY caller of
 // execFileSync in this module — so if it's never invoked, neither is a real
@@ -55,6 +56,22 @@ describe('acquireIntegrationLock / isIntegrationLockLive — self-PID shortcut a
 
     const res = acquireIntegrationLock({ lockPath, pid: process.pid + 1 });
     expect(res).toEqual({ acquired: false, holderPid: process.pid });
+    expect(execFileSyncMock).not.toHaveBeenCalled();
+  });
+
+  it('isIntegrationLockLive: a lock held by our PARENT pid resolves true without ever calling `ps`', () => {
+    // OMN-263 review (pass 3): the topology that actually matters in a real
+    // run. globalSetup writes the vitest MAIN process's pid to the lock, but
+    // per-file afterAll fullCleanup() calls run in the forked WORKER — whose
+    // process.pid differs from the holder, so the self-PID case alone never
+    // fired there and every teardown paid a real `ps` shellout. The worker's
+    // direct parent IS the main process (tinypool forks workers from it), so
+    // holder === process.ppid must resolve true with zero subprocess calls.
+    // This test process is itself a vitest fork worker, making process.ppid
+    // exactly that live vitest-main relationship, not a simulation of it.
+    fs.writeFileSync(lockPath, String(process.ppid), 'utf-8');
+
+    expect(isIntegrationLockLive({ lockPath })).toBe(true);
     expect(execFileSyncMock).not.toHaveBeenCalled();
   });
 });
